OPERATORS IN SQL
An operator is a reserved character or word which is used in a SQL statement to query our database. We use a WHERE clause to query a database using operators.

Operators are needed to specify conditions in a SQL statement. The available operators act as a connector for various conditional statements.
BETWEEN
LIKE
IN
NOT
<
>
<>
>=
<=