Create DATEDIFF & DATEADD FUNCTION
DATEDIFF
DATABASE: AdventureWorks 2014

Table: HumanResources.Employee

To produce a report of all employees with more than ten years of service.

To do this, we’ll use the DATEDIFF report to both display number of years of service and to also to filter out those with less than ten years of service.

DATABASE: AdventureWorks 2014

Table: Production.Product

The sales manager has been talking to the production manager about selling items on the web. They want to produce the bikes as they are ordered. If bikes were ordered today, when could be they be available to ship to the customer?
DATEADD
Using DATEADD in conjunction with DaysToManufacture ,GETDATE(), you can calculate the number of days into the future.