Create CAST FUNCTION
The CAST function converts the value of a column into a specific data type. This helps to ensure that the data types of the values in the expression are compatible.

SYNTAX:

CAST(Expression AS [Datatype])
Expression: It is a valid expression where we want to convert a data type into the SQL.

Datatype: It is a Data Type to which we want to convert the expression.

Length: It is the optional parameter of an integer type. We can use the setting to define the length of any targeted data type. By default, it is 30
DATABASE: AdventureWorks 2014

TABLE: Sales.CurrencyRate

CAST function is used to convert the money data type into integer data type.
Sales.CurrencyRate table is used in this example to show how you can change the data type(money) of the column ''AverageRate''  into integer. 
The ''AverageRate'' column name is changed into new column name ''AvgRate'' and by using CAST function the values of this column is changed into integer value.