Create COALESCE FUNCTION
The COALESCE function accepts a number of arguments and returns the first non-NULL argument. The following illustrates the syntax of the COALESCE function.

The COALESCE function evaluates its arguments from left to right. It stops evaluating until it finds the first non-NULL argument. It means that all the remaining arguments are not evaluated at all
EXAMPLE:


DATABASE: AdventureWorks2014

TABLE: Person.Person